requirements.txt
paho-mqtt
requests
notifiers
-e libs/docker_secrets
